# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "https://app.agm-sa.net/api"
        }
    ],
    "paths": {
...
        "/provider/jobs": {
            "post": {
                "tags": [
                    "ProviderJobs"
                ],
                "summary": "Create new job ",
                "operationId": "dbb7daebfb656c7d777010b304c489b3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "job_title": {
                                        "type": "string"
                                    },
                                    "job_description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "response": "created_successfully",
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/provider/jobs/{id}": {
            "get": {
                "tags": [
                    "ProviderJobs"
                ],
                "summary": "Get Job ",
                "operationId": "f6947b7845e8130cbeaed604135de3ec",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "job": [],
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ProviderJobs"
                ],
                "summary": "Update job ",
                "operationId": "46f7cf66302ace533d4b7833d56d243b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "job_title": {
                                        "type": "string"
                                    },
                                    "job_description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "task": [],
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProviderJobs"
                ],
                "summary": "Delete job ",
                "operationId": "8edafe482d4b2f85bea22295c52666dc",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "response": "deleted_successfully",
                                            "success": true
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}